fix : Добавление проверок в ServiceMpa.java

+++ src/main/java/ru/yandex/practicum/filmorete/service/ServiceMpa.java
@@ -24,8 +24,8 @@ public class ServiceMpa {
 
     public Mpa getSearchId(Integer mpaId) {
         Optional<Mpa> optional = mpaDao.findMpaByRowId(mpaId);
-        if (optional.isPresent()) return optional.get();
-        else throw new ExceptionNotFoundMpaStorage(SERVICE_ERROR_MPA_NOT_IN_MPA_COLLECTIONS);
+        if (optional.isEmpty()) throw new ExceptionNotFoundMpaStorage(SERVICE_ERROR_MPA_NOT_IN_MPA_COLLECTIONS);
+        return optional.get();
     }
 
     public List<Mpa> getAll() {
@@ -33,12 +33,15 @@ public class ServiceMpa {
     }
 
     public void add(@NotNull Mpa mpa) {
-        if (mpa.getId() == null) mpaDao.insert(mpa.getName(), mpa.getDescription());
-        else mpaDao.insert(mpa.getId(), mpa.getName(), mpa.getDescription());
+        Optional<Mpa> optionalMpa = mpaDao.findMpaByRowId(mpa.getId());
+        if (optionalMpa.isEmpty()) throw new ExceptionNotFoundMpaStorage(SERVICE_ERROR_MPA_NOT_IN_MPA_COLLECTIONS);
+        mpaDao.insert(mpa.getName(), mpa.getDescription());
     }
 
     public void update(@NotNull Mpa mpa) {
-        if (mpa.getId() != null) mpaDao.update(mpa.getId(), mpa.getName(), mpa.getDescription());
+        Optional<Mpa> optionalMpa = mpaDao.findMpaByRowId(mpa.getId());
+        if (optionalMpa.isEmpty()) throw new ExceptionNotFoundMpaStorage(SERVICE_ERROR_MPA_NOT_IN_MPA_COLLECTIONS);
+        mpaDao.update(mpa.getId(), mpa.getName(), mpa.getDescription());
     }
 
     public void deleteAll() {
@@ -46,6 +49,8 @@ public class ServiceMpa {
     }
 
     public void deleteSearchId(Integer mpaId) {
+        Optional<Mpa> optionalMpa = mpaDao.findMpaByRowId(mpaId);
+        if (optionalMpa.isEmpty()) throw new ExceptionNotFoundMpaStorage(SERVICE_ERROR_MPA_NOT_IN_MPA_COLLECTIONS);
         mpaDao.deleteByRowId(mpaId);
     }
 }
